fix(dbusservice): Use correct D-Bus type for array-of-dicts
The methods `GetAuthenticationModes` and `SelectAuthenticationMode` use
the type `a{ss}` for parameters that have the type `[]map[string]string`
in the Go code.

However, according the D-Bus specification, `a{ss}` represents an "array
of dict entries", not an "array of dicts" which should be translated to
`aa{ss}`. This can also be observed in the output of
`sudo busctl monitor com.ubuntu.authd.MSEntraID`.

Fixes: ubuntu/authd#628